requirements.txt
Pennylane>=0.36
qiskit>=1.2.0
qiskit-aer>=0.15.1
matplotlib>=3.8.0
torch>=2.2.0
rdkit-pypi>=2022.9.5
numpy>=1.24.0

